queried and obtained firebase details - yay!!

+++ src/main/java/com/supimon/cheflistservice/models/ChefItem.java
@@ -6,9 +6,9 @@ public class ChefItem {
     private String name;
     private String imgUrl;
     private String skills;
-    private int experience;
+    private Long experience;
 
-    public ChefItem(String chefId, String name, String imgUrl, String skills, int experience) {
+    public ChefItem(String chefId, String name, String imgUrl, String skills, Long experience) {
         this.chefId = chefId;
         this.name = name;
         this.imgUrl = imgUrl;
@@ -48,11 +48,11 @@ public class ChefItem {
         this.skills = skills;
     }
 
-    public int getExperience() {
+    public Long getExperience() {
         return experience;
     }
 
-    public void setExperience(int experience) {
+    public void setExperience(Long experience) {
         this.experience = experience;
     }
 }
